+++ include/mockturtle/networks/klut.hpp
@@ -466,7 +466,7 @@ public:
     std::for_each( _storage->nodes.begin(), _storage->nodes.end(), []( auto& n ) { n.data[1].h2 = 0; } );
   }
 
-  uint32_t visited( node const& n ) const
+  auto visited( node const& n ) const
   {
     return _storage->nodes[n].data[1].h2;
   }
@@ -488,10 +488,6 @@ public:
 #pragma endregion
 
 #pragma region General methods
-  void update()
-  {
-  }
-
   auto& events() const
   {
     return *_events;