+++ include/mockturtle/networks/xag.hpp
@@ -1098,10 +1098,6 @@ public:
 #pragma endregion
 
 #pragma region General methods
-  void update()
-  {
-  }
-
   auto& events() const
   {
     return *_events;