+++ include/mockturtle/networks/xmg.hpp
@@ -959,10 +959,6 @@ public:
 #pragma endregion
 
 #pragma region General methods
-  void update()
-  {
-  }
-
   auto& events() const
   {
     return *_events;